% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask}
\alias{ask}
\title{Ask a quiz question}
\usage{
ask(question, answer)
}
\arguments{
\item{question}{Named character string to be used as the question. See details for valid names.}

\item{answer}{Character string to be used as the answer. See details for valid names.}
}
\value{
An object of class \code{iRc_item}.
}
\description{
Creates an object of class \code{\link[iRcotofun:iRc_item-class]{iRc_item}}.
}
\details{
Both parameters must be named character strings (or vectors). Currently these names are supported:
\describe{
   \item{text: }{The value will be pasted as-is to the HTML document.}
   \item{img: }{The value wil be treated as the relative path to an image.}
   \item{html: }{The value must be a XiMpLe.node object.}
}
}
\examples{
q1 <- ask(
  question=list(text="What is Crichton's nick name for Chiana?"),
  answer=list(html=strong("Pip"))
)
q2 <- ask(
  question=list(text="What is Crichton's nick name for his gun?"),
  answer=list(html=strong("Winona"))
)
q3 <- ask(
  question=list(text="What is a famous quote from Rygel?"),
  answer=list(html=strong("Hail, prince of the obvious!"))
)
q4 <- ask(
  question=list(text="What is the Nebari Resistance fighting against?"),
  answer=list(html=strong("The Establishment"))
)
}

