% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\name{category}
\alias{category}
\title{Create a quiz category}
\usage{
category(name, ...)
}
\arguments{
\item{name}{Character string, name of the category.}

\item{...}{Objects of class \code{\link[iRcotofun:iRc_item-class]{iRc_item}}.}
}
\value{
An object of class \code{iRc_category}.
}
\description{
Creates an object of class \code{\link[iRcotofun:iRc_category-class]{iRc_category}}.
}
\examples{
q1 <- ask(
  question=list(text="What is Crichton's nick name for Chiana?"),
  answer=list(html=strong("Pip"))
)
q2 <- ask(
  question=list(text="What is Crichton's nick name for his gun?"),
  answer=list(html=strong("Winona"))
)
q3 <- ask(
  question=list(text="What is a famous quote from Rygel?"),
  answer=list(html=strong("Hail, prince of the obvious!"))
)
q4 <- ask(
  question=list(text="What is the Nebari Resistance fighting against?"),
  answer=list(html=strong("The Establishment"))
)
# now make a category
farscape <- category(name="Farscape", q1, q2, q3, q4)
}

