% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_02_iRc_category.R
\docType{class}
\name{iRc_category,-class}
\alias{iRc_category,-class}
\alias{iRc_category-class}
\title{Class iRc_category}
\description{
Objects of this class represent one category in a iRcotofun game. They are combined to a game
by the function \code{\link[iRcotofun:ircotofun]{ircotofun}}.
To create categories, use the function \code{\link[iRcotofun:category]{category}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character string, name of the category.}

\item{\code{items}}{A list of objects of class \code{\link[iRcotofun:iRc_item-class]{iRc_item}}.}
}}
\keyword{classes}

