% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_01_iRc_item.R
\docType{class}
\name{iRc_item,-class}
\alias{iRc_item,-class}
\alias{iRc_item-class}
\title{Class iRc_item}
\description{
Objects of this class represent one item in a iRcotofun game. They are combined to a category
in objects of class iRc_Category. To freate items,
      use the function \code{\link[iRcotofun:ask]{ask}}.
}
\details{
Both slots must be named character strings (or vectors). Currently these names are supported:
\describe{
   \item{text: }{The value will be pasted as-is to the HTML document.}
   \item{img: }{The value will be treated as the relative path to an image.}
   \item{html: }{The value must be a XiMpLe.node object.}
}
}
\section{Slots}{

\describe{
\item{\code{question}}{Named list to be used as the question. See details for valid names.}

\item{\code{answer}}{Named list to be used as the answer. See details for valid names.}
}}
\keyword{classes}

