% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iRcotofun-package.R
\docType{package}
\name{iRcotofun-package}
\alias{iRcotofun-package}
\title{The iRcotofun Package}
\description{
A Generator for Jeopardy-like Quiz Games (HTML).
}
\details{
\tabular{ll}{
Package: \tab iRcotofun\cr
Type: \tab Package\cr
Version: \tab 0.02-1\cr
Date: \tab 2015-12-13\cr
Depends: \tab R (>= 2.9.0),methods,base64enc,XiMpLe\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://reaktanz.de/?c=hacking&s=iRcotofun\cr
}

Provides some functions and classes to build your own quiz with a few lines of code.
The result is a single HTML file, i.e.,
      the quiz is played using a web browser. The package was
written to have fun at our famous DiagDiff Weihnachtsfeier ;-)
}
\author{
Meik Michalke
}
\keyword{package}

