% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ircotofun.R
\name{ircotofun}
\alias{ircotofun}
\title{Create a quiz}
\usage{
ircotofun(..., points, file = NULL, title = "iRcotofun", sound = NULL,
  css = NULL, overwrite = FALSE, questions = FALSE)
}
\arguments{
\item{...}{Objects of class \code{\link[iRcotofun:iRc_category-class]{iRc_category}}.}

\item{points}{A numeric vector defining the points for each item in
all catgories.}

\item{file}{Character string, path to a file to write to.}

\item{title}{Character string, title of the quiz.}

\item{sound}{Character string,
      name of the sound file to use for background (if available).}

\item{css}{Character string,
      path to a custom CSS file if you don't want to use the default.}

\item{overwrite}{Logical,
      whether existing files should be overwritten. Otherwise an error is thrown.}

\item{questions}{Logical,
      whether the questions should be shown (quiz style) or the answers (jeopardy style, default).}
}
\value{
An object of class \code{XiMpLe.doc}, or (if \code{file} is specified) no visible
   return value.
}
\description{
Creates an object of class \code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}},
which in this case is an HTML document that can be written to a single file
and opened in a web browser to get a quiz.
}
\details{
All categories must have the same number of items, and this number must also
be identical to the values given as \code{points}.
Also, if you use images, make sure that paths are valid.
}
\examples{
q1 <- ask(
  question=list(text="What is Crichton's nick name for Chiana?"),
  answer=list(html=strong("Pip"))
)
q2 <- ask(
  question=list(text="What is Crichton's nick name for his gun?"),
  answer=list(html=strong("Winona"))
)
q3 <- ask(
  question=list(text="What is a famous quote from Rygel?"),
  answer=list(html=strong("Hail, prince of the obvious!"))
)
q4 <- ask(
  question=list(text="What is the Nebari Resistance fighting against?"),
  answer=list(html=strong("The Establishment"))
)

# make a category
farscape <- category(name="Farscape", q1, q2, q3, q4)

# for the sake of demonstration, we'll create a quiz that shows
# the same category four times...
\dontrun{
(output <- tempfile(fileext=".html"))
ircotofun(
  farscape,
  farscape,
  farscape,
  farscape,
  points=c(100,200,300,400),
  file=output,
  title="Best Science Fiction Shows Ever"
)
}
}

