% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titlepage.R
\name{titlepage}
\alias{titlepage}
\title{Generate a title page}
\usage{
titlepage(..., quizfile, file = NULL, title = "iRcotofun", sound = NULL,
  css = NULL, overwrite = FALSE)
}
\arguments{
\item{...}{Objects of class \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}},
      the page content.}

\item{quizfile}{Character string,
      full path to the actual quiz file. The resulting title page will
use a relative link if both files are in the same directory,
      so it's easy to move both afterwards.}

\item{file}{Character string, path to a file to write to.}

\item{title}{Character string, used as the window title.}

\item{sound}{Character string,
      name of the sound file to use for automatic background (if available).}

\item{css}{Character string,
      path to a custom CSS file if you don't want to use the default.}

\item{overwrite}{Logical,
      whether existing files should be overwritten. Otherwise an error is thrown.}
}
\value{
An object of class \code{XiMpLe.doc}, or (if \code{file} is specified) no visible
   return value.
}
\description{
In case you would like to show off with a huge show,
      use this function to generate a title page
that plays some intro music, before you open the actual quiz.
}
\examples{
\dontrun{
# for a logo you'll probably get best results with scalable SVG
logo <- div(
  img(
    attrs=list(
      src=file.path("path", "to", "myQuiz.svg"),
      class="image",
      style="margin-top: 10\%;"
    )
  ),
  attrs=list(class="imagediv")
)

# finally, write the title page
titlepage(
  logo,
  quizfile=file.path("path", "to", "myQuiz.html"),
  file=file.path("path", "to", "index.html"),
  sound=file.path("path", "to", "myQuiz_main_title.ogg"),
  overwrite=TRUE
)
}
}

